Nouvelle version du graphe en utilisant JOGL (version provisoire)

+++ src/graph/Community.java
@@ -1,14 +1,24 @@
 package graph;
 
+/**
+ * Représente une communauté du graphe
+ */
 public class Community {
+
     private final int id;
-    private final double color_r;
-    private final double color_g;
-    private final double color_b;
+    private final float color_r;
+    private final float color_g;
+    private final float color_b;
 
 
-
-    public Community(int id, double r, double g, double b) {
+    /**
+     * Crée une communauté à partir de son identifiant et de ses composantes de couleur
+     * @param id : identifiant de la communauté
+     * @param r : composante rouge de la couleur de la communauté
+     * @param g : composante verte de la couleur de la communauté
+     * @param b : composante bleue de la couleur de la communauté
+     */
+    public Community(int id, float r, float g, float b) {
         this.id = id;
         color_r = r;
         color_g = g;
@@ -16,20 +26,38 @@ public class Community {
     }
 
 
+    /**
+     * @return l'identifiant de la communauté
+     */
     public int getId() {
         return id;
     }
 
-    public double getR() {
+    /**
+     * @return la composante rouge de la couleur de la communauté
+     */
+    public float getR() {
         return color_r;
     }
-    public double getG() {
+
+    /**
+     * @return la composante verte de la couleur de la communauté
+     */
+    public float getG() {
         return color_g;
     }
-    public double getB() {
+
+    /**
+     * @return la composante bleue de la couleur de la communauté
+     */
+    public float getB() {
         return color_b;
     }
 
+
+    /**
+     * @return une représentation textuelle de la communauté
+     */
     public String toString() {
         return "Communauté " + id + " (R : " + String.format("%.2f", color_r) + ", G : " + String.format("%.2f", color_g) + ", B : " + String.format("%.2f", color_b) + ")";
     }
